Nevada/legal_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "nevada"
version = "1.0.0"
description = "Nevada — Agente de Dictamen Jurídico para KYB (Banco PagaTodo)"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.12",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[tool.setuptools.packages.find]
include = ["legal_agent*"]
